+++ src/test/suite/extensionFunctions.test.ts
@@ -1,6 +1,7 @@
 import * as assert from 'assert';
 import * as vscode from 'vscode';
 import {extensionFunctions} from '../../extensionFunctions';
+import { handleExtFilePath } from '../../commandHandler';
 
 suite('Extension Functions Test Suite', () => {
 
@@ -74,5 +75,25 @@ suite('Extension Functions Test Suite', () => {
 		await extensionFunctions.cliExecutionCheck('test-bogus','doesnotexist').then( (returnBool) => {
 			assert.equal(returnBool, false);
 		});
+	});
+
+	test('test to make sure we can get the file path from a whole Uri', async function() {
+		const shouldEndWith = 'vscode-didact/demo/didact-demo.didact.md';
+		const shouldEndWithRaw = 'vscode-didact/master/demo/didact-demo.didact.md';
+
+		const textUri = vscode.Uri.parse('vscode://redhat.vscode-didact?extension=redhat.vscode-didact/demo/didact-demo.didact.md');
+		const rtnUri = extensionFunctions.handleVSCodeDidactUriParsingForPath(textUri);
+		console.log('---------------handleVSCodeDidactUriParsingForPath returned (1) = ' + rtnUri);
+		assert.strictEqual(rtnUri?.fsPath.endsWith(shouldEndWith), true);
+
+		const textUri2 = vscode.Uri.parse('vscode://redhat.vscode-didact?extension=demo/didact-demo.didact.md');
+		const rtnUri2 = extensionFunctions.handleVSCodeDidactUriParsingForPath(textUri2);
+		console.log('---------------handleVSCodeDidactUriParsingForPath returned (2) = ' + rtnUri2);
+		assert.strictEqual(rtnUri2?.fsPath.endsWith(shouldEndWith), true);
+
+		const textUri3 = vscode.Uri.parse('vscode://redhat.vscode-didact?https=raw.githubusercontent.com/redhat-developer/vscode-didact/master/demo/didact-demo.didact.md');
+		const rtnUri3 = extensionFunctions.handleVSCodeDidactUriParsingForPath(textUri3);
+		console.log('---------------handleVSCodeDidactUriParsingForPath returned (3) = ' + rtnUri3);
+		assert.strictEqual(rtnUri3?.fsPath.endsWith(shouldEndWithRaw), true);
 	});
 });